+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # Luma | Framework Component Changelog
 
+## [Unreleased]
+- Update installer to account for planned changes to `package.json` within the Luma project
+
 ## [0.4.1] - 2024-03-02
 - Update dependencies